+++ src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,2 +1,1 @@
 io.github.hee9841.excel.annotation.processor.ExcelAnnotationProcessor
-io.github.hee9841.excel.annotation.processor.ExcelColumnAnnotationProcessor
